Review 1
Made a couple of chnages to the portfolio section: Tied the link to the entire image instead of just the text, killed the hover function on smaller screen sizes so the linear gradient effect is on the image so the user can see the description before they click since mobile devices are just screens and no mouse to hover. P.S: Really have to work on my description,too long lol.